Updating Database File Capabilities, Changed VBIN Processing
VBIN files now read as a tree of SceneNodes and Meshes, making processing more accurate. This came with an overhaul of the entire reading code, so bugs likely exist.

TMD and TDB files can now be read, browsed, and have their contents changed. There are exceptions to this, as not every data type has been accounted for and only one of each file can be loaded for now. The infrastructure to expand on each of these does currently exist.