A Number of Updates
This includes a significant number of changes, not all of which have been documented. Most are focused on Database and Texture modes.

There is now an indicator for the current file mode and the last opened file.

Some code exists for testing changes to VAC files to reverse-engineer in order to convert into .wav files.