Major VBIN Read Overhaul
Redid the entire VBIN reading code to read the file in order (the way files are normally read) instead of searching for specific chunks. A significant amount of the data in the file is still unknown but hopefully reading the file this way will help to track down what they are.