ProgWindow Detangling, File Handling, GUI Update
This updated contains a couple changes for the user:
1. The GUI has been updated and should look more like a "real" program.
There's still quite a ways to go with this, but there are plans for the
future.
2. Multiple files can be loaded at a time. This will allow for
information to be shared across files (like textures to models) down the
road, but for now is just a convenience.

These changes come with significant code updates:
1. The ProgWindow class has been mostly detangled from the other
classes, now just acting as what it is: the program window. The Warpgate
calculator is still tied to this for now, I just haven't able to think
of a clean solution yet.
2. Because of this, it is now the responsibility of each file class to
create their needed GUI elements.

Bugfixes:
Loading an ITF file after a VBIN will no longer crash the program,
thanks to the new file system.

Known issues:
Editing and saving TMD/TDB/BMB/BDB is currently broken. Using versions
<5.0 will work best for them until this is fixed. Opening and browsing
the files still works.
MESH.VBIN files do not read correctly yet, these are still being
researched.