New GUI and Multi-File Loading
This update de-tangles ProgWindow from the file classes, making it the
responsibility of each file class to set up its interface and make
connections. It also allows for multiple files to be loaded at once,
opening the door for cross-file referencing in later updates.

This comes with numerous bugfixes, but also some pretty large bugs.
Database files can't be edited and can't be saved. These will be getting
an update soon to correct them.

This update also corrects an issue with 8bpp ITF files, which had what
were believed to be swizzling issues. The actual problem was that the
palette is written differently to the BMP format than to ITF. This has
now been corrected.