Updating to 0.6.6.0
The dictionary system was completely rewritten to be more accurate for
file reading and writing. This new system has specific data types for
each possible attribute in an item. This also allows for easier
exporting of this information to other file formats, like DAE.

Changes were also made to object and level VBIN models to be more
accurate. The largest of these are that object models have had an
algorithm added that should correct normals and level models now
correctly read and export color data.

There are also a significant number of small, undocumented changes.