0.6.6.6 Correcting Github, Successful Warpgate Loading
Still troubleshooting the github update - it didn't seem to correctly
move the DistanceCalculator files to the Database folders.

Additionally, warpgates now correctly copy to the DistanceCalculator
class. Work still needs to be done to correct the warpgate list that is
initially created from the database file as there are extra warpgates.