0.6.6.7 ITF Bugfixes
Found and fixed a couple bugs with importing and exporting:
Importing an image with a bit depth other than 8bpp would not change the
property byte to reflect the conversion, causing errors on export
attempts.
All ITF exports have had their height and width swapped. For square
textures this was not an issue, for rectangles it obviously was.