Dynamic Mod and Location Loading
Added reading functions that allow the Randomizer to dynamically load
mod and location files from respective folders. The mod file format was
established with SELF's text mod export and then used for the custom
location format. These will make it easier for the community to add new
features to the Randomizer without requiring a patch.

Related plans: Function to create and export custom locations to ensure
the formatting is correct. Add similar function for file replacements.