Minicon Randomization and Writing
With the changes made to DataHandler, Minicon data is now handled
differentlly. The entire list of Minicons is read from the METAGAME
file, then the CREATURE files are read. The data from both of these are
combined to make the final Minicon data that can be used for
randomization of both stats and locations.